PHONEBOOK - Doresa SHUKAJ
My testing skills about PHP&MySQL.
With database,I created a usser acces table and created 3 users. Then, i created a table of countries.
With PHP, I created a page when one of them is the main one that includes 1 title, 2 menus and a block. When the user clicks on the menu,corrospondes at the block.
The pages are:index.php, phonebook.php, login.php, contact.php.
Check out to see how it turned out :)